♻️ - Planner screen header cleanup (#227)
