+++ app/components/index.ts
@@ -15,3 +15,4 @@ export * from "./favorite-routes/favorite-routes"
 export * from "./date-picker-modal"
 export * from "./list"
 export * from "./sheets"
+export * from "./chip"
